requirements.txt
Flask==3.0.3
openai==1.37.1
langchain==0.2.11
langchain-community==0.2.10
python-dotenv==1.0.1
flask-cors==3.0.10
Werkzeug>=3.0.0
PyPDF2
Pillow
python-docx
pytesseract